backend/requirements.txt
fastapi>=0.115.9
uvicorn>=0.27.1
pydantic>=2.8.0
pydantic-settings>=2.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
email-validator>=2.1.0.post1
groq>=0.4.0 